[TA2299] Fix replica registration bug
On branch fix-replica-registration-bug
Changes to be committed:
	modified:   ../backend/remote/remote.go
	new file:   ../backend/remote/signal.go
	new file:   ../backend/remote/signal_debug.go
	modified:   ../controller/control.go
	modified:   ../controller/rest/model.go
	modified:   ../frontend/gotgt/frontend.go
	modified:   ../replica/rest/replica.go
	new file:   ../sync/add_replica.go
	new file:   ../sync/add_replica_debug.go
	modified:   ../sync/sync.go

1. Why is this change necessary?
-  Controller was trying to connect with the first registered replica only, when
   that replica gets crashed before getting start signal.
-  Controller was not caching the error in case when 'start' signal was not sent
   successfully.
-  Need for better and descriptive logs while registration of replica.

2. How does it address the issue?
-  This commit catch the error after signalling replicas and sets
   c.MaxRevReplica = "" and also deletes the entry for this replica from map.
-  Adds descriptive logs.

3. What side effects does this change have?
-  None
